pyproject.toml
[project]
name = "introduction-to-mcp-with-sql-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.49.0",
    "docker>=7.1.0",
    "instructor[claude]>=1.7.3",
    "loguru>=0.7.3",
    "mcp[cli]>=1.3.0",
    "openai==1.70.0",
    "pydantic>=2.10.6",
    "python-dotenv>=1.0.1",
    "rich>=13.9.4",
]

[dependency-groups]
dev = ["types-docker>=7.1.0.20241229"]
